LayoutTests/media/video-play-pause-events-expected.txt

Test that calling play() and pause() triggers async play, timeupdate and pause events.

RUN(video.play())
RUN(video.pause())
SCRIPT DONE
EVENT(play)
EXPECTED (video.paused == 'true') OK
EVENT(timeupdate)
EXPECTED (video.paused == 'true') OK
EVENT(pause)
EXPECTED (video.paused == 'true') OK
END OF TEST
